# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SubjectPriority = typeof SubjectPriority[keyof typeof SubjectPriority];


export const SubjectPriority = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface Subject {
  id: number;
  name: string;
  /**
     * ISO date string (YYYY-MM-DD) or null
     * @nullable
     */
  examDate?: string | null;
  priority: SubjectPriority;
  hoursPerDay: number;
  completed: boolean;
  /**
     * Comma-separated topics
     * @nullable
     */
  topics?: string | null;
  createdAt: string;
}

export type SubjectInputPriority = typeof SubjectInputPriority[keyof typeof SubjectInputPriority];


export const SubjectInputPriority = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface SubjectInput {
  /** @minLength 1 */
  name: string;
  /**
     * ISO date string (YYYY-MM-DD) — optional
     * @nullable
     */
  examDate?: string | null;
  priority: SubjectInputPriority;
  /**
     * @minimum 0.5
     * @maximum 24
     */
  hoursPerDay: number;
  /** Comma-separated topics */
  topics?: string;
}

export type SubjectUpdatePriority = typeof SubjectUpdatePriority[keyof typeof SubjectUpdatePriority];


export const SubjectUpdatePriority = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface SubjectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  examDate?: string | null;
  priority?: SubjectUpdatePriority;
  /**
     * @minimum 0.5
     * @maximum 24
     */
  hoursPerDay?: number;
  topics?: string;
}

export type StudySessionPriority = typeof StudySessionPriority[keyof typeof StudySessionPriority];


export const StudySessionPriority = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface StudySession {
  id: number;
  subjectId: number;
  subjectName: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  hours: number;
  completed: boolean;
  priority: StudySessionPriority;
  /** @nullable */
  topic?: string | null;
}

export interface TimetableConfig {
  /** ISO date string to start timetable from */
  startDate: string;
}

export type UpcomingExamPriority = typeof UpcomingExamPriority[keyof typeof UpcomingExamPriority];


export const UpcomingExamPriority = {
  High: 'High',
  Medium: 'Medium',
  Low: 'Low',
} as const;

export interface UpcomingExam {
  id: number;
  name: string;
  examDate: string;
  priority: UpcomingExamPriority;
  daysLeft: number;
}

export interface DashboardStats {
  totalSubjects: number;
  completedSubjects: number;
  pendingSubjects: number;
  totalSessions: number;
  completedSessions: number;
  progressPercentage: number;
  upcomingExams: UpcomingExam[];
}

export interface Note {
  id: number;
  title: string;
  body: string;
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  /** @minLength 1 */
  title: string;
  body?: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListSubjectsParams = {
/**
 * Search term for filtering subjects
 */
search?: string;
};