# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { UpcomingExam } from './upcomingExam';

export interface DashboardStats {
  totalSubjects: number;
  completedSubjects: number;
  pendingSubjects: number;
  totalSessions: number;
  completedSessions: number;
  progressPercentage: number;
  upcomingExams: UpcomingExam[];
}