# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { StudySessionPriority } from './studySessionPriority';

export interface StudySession {
  id: number;
  subjectId: number;
  subjectName: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  hours: number;
  completed: boolean;
  priority: StudySessionPriority;
  /** @nullable */
  topic?: string | null;
}