# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectUpdatePriority } from './subjectUpdatePriority';

export interface SubjectUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  examDate?: string | null;
  priority?: SubjectUpdatePriority;
  /**
     * @minimum 0.5
     * @maximum 24
     */
  hoursPerDay?: number;
  topics?: string;
}