# lib/api-zod/src/generated/types/timetableConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Study Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TimetableConfig {
  /** ISO date string to start timetable from */
  startDate: string;
}